requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
line-bot-sdk==3.11.0
google-generativeai==0.8.3
notion-client==2.2.1
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
